--- tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/antag_tips.tsx
@@ -1,9 +1,0 @@
-import { CheckboxInput, type FeatureToggle } from '../base';
-
-export const antag_tips: FeatureToggle = {
-  name: 'Antagonist tips',
-  category: 'GAMEPLAY',
-  description:
-    'Gives a basic explanation on how to play an antagonist upon obtaining one.',
-  component: CheckboxInput,
-};
