+++ tgui/packages/tgui/interfaces/PreferencesMenu/preferences/features/game_preferences/fulp_settings.tsx
@@ -1,3 +1,13 @@
+import { CheckboxInput, type FeatureToggle } from '../base';
+
+export const antag_tips: FeatureToggle = {
+  name: 'Antagonist tips',
+  category: 'GAMEPLAY',
+  description:
+    'Gives a basic explanation on how to play an antagonist upon obtaining one.',
+  component: CheckboxInput,
+};
+
 import { Box, TextArea } from 'tgui-core/components';
 import type { Feature, FeatureValueProps } from '../base';
